pyproject.toml
[project]
name = "project-data-football"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "databricks-sql-connector>=4.2.5",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
